Shrink `TypeId` back to 16-bytes
This commit preserves the 16-byte size of `TypeId` from `main` by
shrinking two fields back from `usize` to `u32`. The `type_size` field
was updated everywhere to use `u32` instead of `usize` since it never
exceeds 100k anyway and doesn't need `usize` on 64-bit platforms. The
`unique_id` integer was additionally shrunk to 32-bit as modules won't
be able to otherwise define 4B types.

Additionally an assertion is added in this commit to ensure that the
size of this structure stays relatively small.

+++ crates/wasmparser/src/validator.rs
@@ -74,7 +74,7 @@ fn check_max(cur_len: usize, amt_added: u32, max: usize, desc: &str, offset: usi
     Ok(())
 }
 
-fn combine_type_sizes(a: usize, b: usize, offset: usize) -> Result<usize> {
+fn combine_type_sizes(a: u32, b: u32, offset: usize) -> Result<u32> {
     match a.checked_add(b) {
         Some(sum) if sum < MAX_WASM_TYPE_SIZE => Ok(sum),
         _ => Err(format_err!(
